src/app/app.component.ts
import { Component, OnInit } from '@angular/core';
import { BehaviorSubject } from 'rxjs';

import { LoaderService } from './services/loader/loader.service';

@Component({
  selector: 'app-root',
  templateUrl: './app.component.html',
  styleUrls: ['./app.component.css']
})
export class AppComponent implements OnInit {
  public showLoader: boolean;

  constructor(
    private loaderService: LoaderService) {
  }

  ngOnInit() {
    this.loaderService.status.subscribe((val) => {
      this.showLoader = val;
    });
  }
}
